backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
yt-dlp>=2023.10.13
websockets>=10.0
pydantic>=1.8.0
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.4
python-dotenv>=0.19.1
email-validator>=2.0.0
pydantic-settings>=2.0.0 